Final bug fixed (hopefully...) and cleaning up
All changes except in word2vec.py are meaningless - pyc files were created because I ran the code from the repo folder and forgot to delete them. 

The init.py was added by me to try and get the code to work on PyCharm, but it shouldn't be there for te final submission, so I removed it.